refactor(isa): extract Memory and RegisterFile into ISA-agnostic isa/ core
Move the two fully ISA-agnostic value types out of the mips namespace into a
new isa:: namespace under include/isa/, the first step toward supporting a
second ISA backend (RISC-V) alongside MIPS.

Memory (byte-addressable little-endian RAM) and RegisterFile (32x32 GPRs with
register 0 hardwired to zero) are common to both MIPS and RV32I, so they now
live in include/isa/{memory,registers}.h and are defined in the isa namespace.
The MIPS O32 ABI mnemonic table (register_abi_name) is MIPS-specific and stays
in mips.

mips/memory.h and mips/registers.h keep `using isa::Memory` / `using
isa::RegisterFile` shims so all existing mips:: callers compile unchanged. No
behavior change; core-only and full debug builds green, all tests pass.
